refactor(store): move state to common lib
