+++ apps/crow-storm/src/app/components/PrimarySearchBar.tsx
@@ -7,15 +7,15 @@ import { createStyles, fade, makeStyles, Theme } from '@material-ui/core/styles'
 import Toolbar from '@material-ui/core/Toolbar'
 import Tooltip from '@material-ui/core/Tooltip'
 import Typography from '@material-ui/core/Typography'
+import Brightness2Icon from '@material-ui/icons/Brightness2'
 import MoreIcon from '@material-ui/icons/MoreVert'
 import SearchIcon from '@material-ui/icons/Search'
+import SettingsIcon from '@material-ui/icons/Settings'
+import { settingsActions } from '@waffle-charm/store'
 import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Link, useHistory, useLocation } from 'react-router-dom'
-import SettingsIcon from '@material-ui/icons/Settings'
 import { useDispatch } from 'react-redux'
-import { settingsActions } from '../+store/features/settings.slice'
-import Brightness2Icon from '@material-ui/icons/Brightness2'
+import { Link, useHistory, useLocation } from 'react-router-dom'
 
 export const SUMMONER_NAME_KEY = 'summonerName'
