+++ jest.config.js
@@ -8,5 +8,6 @@ module.exports = {
     '<rootDir>/libs/material',
     '<rootDir>/libs/summoner',
     '<rootDir>/libs/testing-utils',
+    '<rootDir>/libs/store',
   ],
 }
