+++ libs/store/README.md
@@ -1,0 +1,7 @@
+# store
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test store` to execute the unit tests via [Jest](https://jestjs.io).
