+++ libs/store/src/lib/lol-version.slice.ts
