+++ nx.json
@@ -49,6 +49,9 @@
     },
     "testing-utils": {
       "tags": []
+    },
+    "store": {
+      "tags": ["redux"]
     }
   }
 }
