+++ tsconfig.base.json
@@ -21,7 +21,8 @@
       "@waffle-charm/mastery": ["libs/mastery/src/index.ts"],
       "@waffle-charm/material": ["libs/material/src/index.ts"],
       "@waffle-charm/summoner": ["libs/summoner/src/index.ts"],
-      "@waffle-charm/testing-utils": ["libs/testing-utils/src/index.ts"]
+      "@waffle-charm/testing-utils": ["libs/testing-utils/src/index.ts"],
+      "@waffle-charm/store": ["libs/store/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
